navigates to BasicSearchListPage with table of courses when search is pressed
