+++ frontend/src/main/pages/BasicCourseSearchPage.js
@@ -2,19 +2,12 @@ import BasicLayout from "main/layouts/BasicLayout/BasicLayout";
 
 import BasicCourseSearchForm from "main/components/BasicCourseSearch/BasicCourseSearchForm";
 import CoursesTable from "main/components/Courses/CoursesTable"
-
+import { Navigate } from 'react-router-dom'
 import { useState } from "react";
 import { useBackendMutation } from "main/utils/useBackend";
 
 
 const Home = () => {
-
-    // every function that starts with "use" is a hook
-    // e.g. useState, useSWR, useAuth0
-
-    // courseJSON is the variable for the state
-    // setCourseJSON is the setter
-    // the parameter to useState is the initial value of the state
 
     const initialCourseJSON = {
         "pageNumber": 1,
@@ -23,7 +16,6 @@ const Home = () => {
         "classes": []
     };
 
-    // courseId, title, sectionNumber, instructor, enroll code, units, total enrolled students, max enrolled
     const [courseJSON, setCourseJSON] = useState(initialCourseJSON);
 
     const objectToAxiosParams = (query) => ({
@@ -39,10 +31,21 @@ const Home = () => {
 
     const onSuccess = (courses) => { return courses };
 
-    const mutation = useBackendMutation(objectToAxiosParams, {onSuccess}, []);
+    const mutation = useBackendMutation(
+        objectToAxiosParams,
+         {onSuccess},
+        
+         ["api/curriculum/curriculum"]
+         );
+
+    const { isSuccess } = mutation
 
     async function fetchBasicCourseJSON(event, query) {
       mutation.mutate(query);
+    }
+
+    if (isSuccess) {
+        return <Navigate to="/basiccoursesearch/index" />
     }
 
     return (
